+++ ddlx.sql
@@ -2885,10 +2885,11 @@ CREATE OR REPLACE FUNCTION ddlx_drop(oid,ddlx_options text[] default '{}')
        when obj.sql_kind = 'SEQUENCE'
        then format(E'DROP %s IF EXISTS %s;\n',obj.sql_kind, obj.sql_identifier)
        when obj.sql_kind is not null
-       then format(E'DROP %s %s%s;\n',
+       then format(E'DROP %s %s%s;%s\n',
                    obj.sql_kind, 
                    case when 'ie' ilike any($2) then 'IF EXISTS ' end,
-                   obj.sql_identifier
+                   obj.sql_identifier,
+		   case when obj.sql_kind = 'TABLE' then ' -- ATTENTION!' end
                    )
        else format(E'-- DROP UNIDENTIFIED OBJECT: %s\n',text($1))
       end